// src/app/model/childeren.ts
export class Children{
    
    _cActive: boolean;
    _cTitle: string;
    _cName: string;
    _cNic: string;
    _cDob: string;
    _cAge: number;
    _cCibc: boolean;
    _cHrbic: boolean;
    _cHrbfc: boolean;
    _cSuhrbc: boolean;
    _cShcbfc: boolean;
    _cHbc: boolean;
    _cHcbc : boolean;

    _cHrbc: boolean;
    
}